phase3/requirements.txt
# Phase 1 deps
anthropic>=0.25.0
openai-whisper>=20231117
pyttsx3>=2.90
pyaudio>=0.2.13
keyboard>=0.13.5
numpy>=1.24.0
colorama>=0.4.6

# Phase 2 deps
chromadb>=0.4.0
sentence-transformers>=2.2.0
pyautogui>=0.9.54
pillow>=10.0.0
playwright>=1.40.0

# Phase 3 additions
langgraph>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
tavily-python>=0.3.0
apscheduler>=3.10.0
requests>=2.31.0
